app/requirements.txt
Django==4.0.4
gunicorn==20.1.0
djangorestframework==3.13.1
django-filter==2.4.0
django-taggit==1.3.0
django-reset-migrations==0.4.0
django-simple-cookie-consent==0.1.1
django-cleanup==5.1.0
markdown==3.3.3
six==1.15.0
Pillow==2.2.1
sweetify==2.0.0
psycopg2-binary==2.8.5
djoser==2.1.0
Sphinx==5.1.1
django4-docs==0.4.0
sphinx-rtd-theme==1.0.0